changed environments and started project
