+++ config/database.yml
@@ -23,36 +23,19 @@ default: &default
   collation: utf8mb4_unicode_ci
   reconnect: true
 
-development: &DEVELOPMENT
-  <<: *default
-
-v3_development:
-  <<: *default
-
-aws_db:
-  <<: *default
-  host: aws-dev-db.co33oyzjqasf.us-east-1.rds.amazonaws.com
-  username: <%= Rails.application.credentials[:development][:database_username] %>
-  password: <%= Rails.application.credentials[:development][:database_password] %>
-
-local:
+development:
   <<: *default
   host: 127.0.0.1
   username: travis
   password: ""
 
-local_dev:
+dev:
   <<: *default
 
 stage:
   <<: *default
 
-v3_stage:
-  <<: *default
-
 production:
   <<: *default
 
-v3_production:
-  <<: *default
 <% end %>
