+++ config/environments/development.rb
@@ -1,84 +1,64 @@
-require "active_support/core_ext/integer/time"
-
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
+  # this craziness is to get logging both to console and to log file just like normal development
+  # normal_logger = ActiveSupport::Logger.new("log/#{Rails.env}.log")
+  # console_logger = ActiveSupport::Logger.new(STDOUT)
+  # combined_logger = console_logger.extend(ActiveSupport::Logger.broadcast(normal_logger))
+  #
+  # combined_logger.formatter = config.log_formatter
+  # config.logger = ActiveSupport::TaggedLogging.new(combined_logger)
 
-  # In the development environment your application's code is reloaded any time
-  # it changes. This slows down response time but is perfect for development
+  # this is so that we can still see output to console, otherwise it gets turned off for some reason with this environment and webrick
+  # config.middleware.insert_before(Rails::Rack::Logger, Rails::Rack::LogTailer)
+  config.log_level = :debug
+
+  # In the development environment your application's code is reloaded on
+  # every request. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
+  config.cache_classes = false
 
-  # Dryad's development environment is a little more stable than Rails assumes;
-  # we use the local environment for truly quick-changing development work. Settings in this
-  # file are a bit different from the defaults for that reason.
-  
-  config.cache_classes = true
+  # Do not eager load code on boot.
+  config.eager_load = false
 
-  # Eager load code on boot.
-  config.eager_load = true
+  # Show full error reports and disable caching.
+  config.consider_all_requests_local       = true
+  config.action_controller.perform_caching = false
 
-  # Show full error reports.
-  config.consider_all_requests_local = true
-
-  # Enable/disable caching. By default caching is disabled.
-  # Run rails dev:cache to toggle caching.
-  if Rails.root.join('tmp', 'caching-dev.txt').exist?
-    config.action_controller.perform_caching = true
-    config.action_controller.enable_fragment_cache_logging = true
-
-    config.cache_store = :memory_store
-    config.public_file_server.headers = {
-      'Cache-Control' => "public, max-age=#{2.days.to_i}"
-    }
-  else
-    config.action_controller.perform_caching = false
-
-    config.cache_store = :null_store
-  end
-
-  # Mailer
-  config.action_mailer.perform_caching = false
-  config.action_mailer.delivery_method = :sendmail
-  config.action_mailer.perform_deliveries = true
-  config.action_mailer.raise_delivery_errors = true
-  config.action_mailer.default_url_options = { host: 'v3-dev.datadryad.org' }
+  # Don't care if the mailer can't send.
+  config.action_mailer.raise_delivery_errors = false
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
-  # Raise exceptions for disallowed deprecations.
-  config.active_support.disallowed_deprecation = :raise
-
-  # Tell Active Support which deprecation messages to disallow.
-  config.active_support.disallowed_deprecation_warnings = []
-
   # Raise an error on page load if there are pending migrations.
   config.active_record.migration_error = :page_load
 
-  # Highlight code that triggered database queries in logs.
-  config.active_record.verbose_query_logs = true
+  # Store uploaded files on the local file system (see config/storage.yml for options)
+  # config.active_storage.service = :local
 
   # Debug mode disables concatenation and preprocessing of assets.
   # This option may cause significant delays in view rendering with a large
   # number of complex assets.
   config.assets.debug = true
 
-  # Suppress logger output for asset requests.
-  config.assets.quiet = true
+  # Asset digests allow you to set far-future HTTP expiration dates on all assets,
+  # yet still be able to expire them through the digest params.
+  config.assets.digest = true
 
-  # Raises error for missing translations.
-  # config.i18n.raise_on_missing_translations = true
+  # Adds additional error checking when serving assets at runtime.
+  # Checks for improperly declared sprockets dependencies.
+  # Raises helpful error messages.
+  config.assets.raise_runtime_errors = true
 
-  # Annotate rendered view with file names.
-  # config.action_view.annotate_rendered_view_with_filenames = true
+  # Raises error for missing translations
+  # config.action_view.raise_on_missing_translations = true
 
-  # Use an evented file watcher to asynchronously detect changes in source code,
-  # routes, locales, etc. This feature depends on the listen gem.
-  config.file_watcher = ActiveSupport::EventedFileUpdateChecker
+  # Do not email users
+  config.action_mailer.raise_delivery_errors = true
+  config.action_mailer.delivery_method = :letter_opener
+  config.action_mailer.perform_deliveries = true
 
-  # Uncomment if you wish to allow Action Cable access from any origin.
-  # config.action_cable.disable_request_forgery_protection = true
-
-  config.hosts << 'v3-dev.datadryad.org'
-  Rails.application.default_url_options = { host: 'v3-dev.datadryad.org' }
-
+  Rails.application.default_url_options = { host: 'localhost', port: 3000}
+  Rack::Attack.enabled = false
+  config.hosts.clear
 end
