+++ config/initializers/bullet.rb
@@ -1,4 +1,4 @@
-if Rails.env.local?
+if Rails.env.development?
   Rails.application.configure do
     config.after_initialize do
       Bullet.enable        = true
