+++ config/webpacker.yml
@@ -7,15 +7,13 @@ default: &default
   public_output_path: packs
   cache_path: tmp/webpacker
   webpack_compile_output: true
-
   # Additional paths webpack should look up modules
   # ['app/assets', 'engine/foo/app/assets']
   additional_paths: []
-
   # Reload manifest.json on all requests so we reload latest compiled packs
   cache_manifest: false
 
-development: &development
+dev:
   <<: *default
   compile: true
 
@@ -46,43 +44,26 @@ development: &development
     watch_options:
       ignored: '**/node_modules/**'
 
-v3_development:
-  <<: *development
-
-local_dev:
-  <<: *development
-
 test:
   <<: *default
   compile: true
-
   # Compile test packs to a separate directory
   public_output_path: packs-test
 
-local:
+development:
   <<: *default
   compile: true
 
-stage: &STAGE
+stage:
   <<: *default
-
   # Production depends on precompilation of packs prior to booting for performance.
   compile: false
-
   # Cache manifest.json for performance
   cache_manifest: true
 
-v3_stage:
-  <<: *STAGE
-
-production: &PRODUCTION
+production:
   <<: *default
-
   # Production depends on precompilation of packs prior to booting for performance.
   compile: false
-
   # Cache manifest.json for performance
   cache_manifest: true
-
-v3_production:
-  <<: *PRODUCTION
