+++ db/data/20241028164918_add_peer_review_custom_text_for_acs.rb
@@ -2,7 +2,7 @@
 
 class AddPeerReviewCustomTextForAcs < ActiveRecord::Migration[7.0]
   def up
-    return if Rails.env.v3_production?
+    return if Rails.env.production?
 
     StashEngine::JournalOrganization.where(name: 'American Chemical Society')
                                     .first
