+++ lib/stash/compressed/zip_info.rb
@@ -1,6 +1,6 @@
 require 'http'
 require 'stringio'
-require 'charlock_holmes/string'
+# require 'charlock_holmes/string'
 require 'zip'
 require 'down/http'
 require 'open3'
